requirements.txt
langchain==0.3.13
langchain-community==0.3.13
langchain-ollama==0.2.2
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
ollama==0.4.4
openai==1.58.1
python-dotenv==1.0.1
pypdf==5.1.0
faiss-cpu==1.9.0
